Extract date helpers
